+++ CHANGES.md
@@ -1,3 +1,9 @@
+## HEAD
+
+- Renamed `handleResponse` to `onResponse`
+
+[HEAD]: https://github.com/mjackson/http-client/compare/v4.0.0-0...HEAD
+
 ## [4.0.0-0]
 > Apr 27, 2016